Add Query base object
