+++ lib/atomic_json.rb
@@ -5,6 +5,7 @@ require 'active_record/connection_adapters/postgresql_adapter'
 require 'atomic_json/version'
 require 'atomic_json/updater'
 require 'atomic_json/query'
+require 'atomic_json/single_update'
 require 'atomic_json/type_validation'
