+++ lib/atomic_json/query.rb
@@ -20,6 +20,8 @@ module AtomicJson
     attr_reader :record, :jsonb_field, :connection, :options
     attr_accessor :query_string
 
+    delegate :quote_column_name, :quote_table_name, :quote, to: :connection
+
     def initialize(record, jsonb_field, options = {})
       @connection = ActiveRecord::Base.connection
       @record = record
@@ -27,13 +29,8 @@ module AtomicJson
       @options = options.reverse_merge!(DEFAULT_OPTIONS)
     end
 
-    def build(attributes)
-      case attributes.keys.count
-      when 0 then raise_attributes_missing
-      when 1 then single_update_query(attributes)
-      else multi_update_query(attributes)
-      end
-      self
+    def build(_attributes)
+      raise NotImplementedError
     end
 
     def execute!
@@ -46,47 +43,6 @@ module AtomicJson
 
       def raise_attributes_missing
         raise QueryError, 'You need at least one JSONB field to create/update'
-      end
-
-      def single_update_query(attributes)
-        keys, value = keys_and_value(attributes)
-
-        self.query_string = build_query(keys, value)
-      end
-
-      def build_query(keys, value)
-        <<~SQL
-          UPDATE #{connection.quote_table_name(record.class.table_name)}
-          SET #{connection.quote_column_name(jsonb_field)} = jsonb_set(
-                #{connection.quote_column_name(jsonb_field)},
-                #{jsonb_quote_keys(keys)},
-                #{jsonb_quote_value(value)},
-                #{options[:create_missing]}
-              )
-          WHERE id = #{connection.quote(record.id)};
-        SQL
-      end
-
-      ##
-      # Traverse the attributes hash, aggregating all hash keys into an
-      # array and keep the last value
-      def keys_and_value(attributes)
-        keys = []
-        if options[:nested]
-          val = loop do
-            key, val = attributes.flatten
-            keys << key.to_s
-            if val.is_a?(Hash)
-              attributes = val
-            else
-              break val
-            end
-          end
-        else
-          keys << attributes.keys.first
-          val = attributes.values.first
-        end
-        [keys, val]
       end
 
   end
